+++ streams/src/index.ts
@@ -1,17 +1,21 @@
 
-export { StreamEvent, StreamItem, StreamDone,
+export { StreamEvent, StreamItem, StreamDone, StreamFail,
+    StreamLogInfo, StreamLogWarn, StreamLogError, StreamHint,
+    EventType,
     c_done, c_item, c_fail,
     c_log_error, c_log_info, c_log_warn,
-    c_hint, c_hint_list as c_result_list, c_hint_single_item as c_result_single_item,
+    c_hint, c_hint_list, c_hint_single_item,
     } from './EventType';
-export { Stream, EventReceiver } from './Stream';
+export { Stream, EventReceiver, EventReceiverCallback, LooseEventReceiver,
+    CallbacksListener, ProtocolError, UsageError } from './Stream';
 export { formatStreamEvent, eventTypeToString } from './formatStreamEvent';
 export { BackpressureStop, exceptionIsBackpressureStop } from './BackpressureStop'
 export { StreamDispatcher } from './StreamDispatcher';
 export { randomHex, randomAlpha } from './randomHex';
 export { ErrorDetails, ErrorWithDetails, captureError, errorAsStreamEvent,
-    startGlobalErrorListener, recordUnhandledError, errorDetailsToString } from './Errors';
-export { StreamProtocolValidator } from './StreamProtocolValidator'
+    toException, startGlobalErrorListener, recordUnhandledError,
+    errorDetailsToString } from './Errors';
+export { StreamProtocolValidator, wrapStreamInValidator } from './StreamProtocolValidator'
 export { dynamicOutputToStream, callbackToStream } from './dynamicOutputToStream';
 export { callbackBasedIterator } from './callbackBasedIterator';
 export { toConsoleLog } from './toConsoleLog';